Account for various hubot-auth install locations
Using require instead of a simple path allows searching for hubot-auth at various levels, so whether it is installed within hubot-heroku or in the top level node_modules directory, it will be found. The current implementation creates an error for those with hubot-auth installed in the top level node_modules directory.